Moved startOf/endOf/truncatedTo/next/previous operators to one place
