+++ core/src/commonMain/kotlin/io/islandtime/Time.kt
@@ -247,33 +247,6 @@ class Time(
         nanosecond: Int = this.nanosecond
     ) = Time(hour, minute, second, nanosecond)
 
-    /**
-     * Return a copy of this time, truncated to the [hour] value. All smaller components will be replaced with zero.
-     */
-    fun truncatedToHours() = copy(minute = 0, second = 0, nanosecond = 0)
-
-    /**
-     * Return a copy of this time, truncated to the [minute] value. All smaller components will be replaced with zero.
-     */
-    fun truncatedToMinutes() = copy(second = 0, nanosecond = 0)
-
-    /**
-     * Return a copy of this time, truncated to the [second] value. All smaller components will be replaced with zero.
-     */
-    fun truncatedToSeconds() = copy(nanosecond = 0)
-
-    /**
-     * Return a copy of this time with the [nanosecond] value truncated to milliseconds.
-     */
-    fun truncatedToMilliseconds() =
-        copy(nanosecond = this.nanosecond / NANOSECONDS_PER_MILLISECOND * NANOSECONDS_PER_MILLISECOND)
-
-    /**
-     * Return a copy of this time with the [nanosecond] value truncated to microseconds.
-     */
-    fun truncatedToMicroseconds() =
-        copy(nanosecond = this.nanosecond / NANOSECONDS_PER_MICROSECOND * NANOSECONDS_PER_MICROSECOND)
-
     companion object {
         val MIN = Time(0, 0)
         val MAX = Time(23, 59, 59, 999_999_999)
