+++ core/src/commonMain/kotlin/io/islandtime/ranges/ZonedDateTimeInterval.kt
@@ -2,6 +2,8 @@ package io.islandtime.ranges
 
 import io.islandtime.*
 import io.islandtime.measures.*
+import io.islandtime.operators.endOfDayAt
+import io.islandtime.operators.startOfDayAt
 import io.islandtime.parser.*
 import io.islandtime.parser.expectingGroupCount
 import io.islandtime.parser.throwParserFieldResolutionException
