+++ core/src/commonTest/kotlin/io/islandtime/DateTimeTest.kt
@@ -22,28 +22,6 @@ class DateTimeTest : AbstractIslandTimeTest() {
     }
 
     @Test
-    fun `Date_startOfDay returns the DateTime at midnight of same day`() {
-        assertEquals(
-            DateTime(
-                Date(2019, Month.JULY, 1),
-                Time.MIDNIGHT
-            ),
-            Date(2019, Month.JULY, 1).startOfDay
-        )
-    }
-
-    @Test
-    fun `Date_endOfDay returns the DateTime just before the end of the same day`() {
-        assertEquals(
-            DateTime(
-                Date(2019, Month.JULY, 1),
-                Time(23, 59, 59, 999_999_999)
-            ),
-            Date(2019, Month.JULY, 1).endOfDay
-        )
-    }
-
-    @Test
     fun `copy() returns a new DateTime replacing the desired values`() {
         assertEquals(
             DateTime(2018, Month.MAY, 8, 12, 0),
