+++ core/src/commonTest/kotlin/io/islandtime/TimeTest.kt
@@ -123,46 +123,6 @@ class TimeTest : AbstractIslandTimeTest() {
     }
 
     @Test
-    fun `truncatedToHours() removes components smaller than hours`() {
-        assertEquals(
-            Time(1, 0),
-            Time(1, 2, 3, 4).truncatedToHours()
-        )
-    }
-
-    @Test
-    fun `truncatedToMinutes() removes components smaller than minutes`() {
-        assertEquals(
-            Time(1, 2),
-            Time(1, 2, 3, 4).truncatedToMinutes()
-        )
-    }
-
-    @Test
-    fun `truncatedToSeconds() removes components smaller than seconds`() {
-        assertEquals(
-            Time(1, 2, 3),
-            Time(1, 2, 3, 4).truncatedToSeconds()
-        )
-    }
-
-    @Test
-    fun `truncatedToMilliseconds() removes components smaller than milliseconds`() {
-        assertEquals(
-            Time(1, 2, 3, 444_000_000),
-            Time(1, 2, 3, 444_555_666).truncatedToMilliseconds()
-        )
-    }
-
-    @Test
-    fun `truncatedToMicroseconds() removes components smaller than microseconds`() {
-        assertEquals(
-            Time(1, 2, 3, 444_555_000),
-            Time(1, 2, 3, 444_555_666).truncatedToMicroseconds()
-        )
-    }
-
-    @Test
     fun `adding a duration of zero doesn't affect the time`() {
         val time = Time(1, 1, 1, 1)
         assertEquals(time, time + Duration.ZERO)
