+++ core/src/commonTest/kotlin/io/islandtime/ZonedDateTimeTest.kt
@@ -231,32 +231,6 @@ class ZonedDateTimeTest : AbstractIslandTimeTest() {
     }
 
     @Test
-    fun `Date_startOfDayAt() creates a ZonedDateTime at the start of the day in a particular time zone`() {
-        assertEquals(
-            ZonedDateTime(
-                DateTime(2019, 5, 20, 0, 0),
-                nyZone
-            ),
-            Date(2019, 5, 20).startOfDayAt(nyZone)
-        )
-
-        // TODO: Add tests where transitions occur during midnight
-    }
-
-    @Test
-    fun `Date_endOfDayAt() creates a ZonedDateTime at the end of the day in a particular time zone`() {
-        assertEquals(
-            ZonedDateTime(
-                DateTime(2019, 5, 20, 23, 59, 59, 999_999_999),
-                nyZone
-            ),
-            Date(2019, 5, 20).endOfDayAt(nyZone)
-        )
-
-        // TODO: Add tests where transitions occur during midnight
-    }
-
-    @Test
     fun `copy() ignores changes to the offset if it isn't valid for the time zone`() {
         assertEquals(
             ZonedDateTime.create(
