+++ core/src/commonTest/kotlin/io/islandtime/ranges/ZonedDateTimeIntervalTest.kt
@@ -3,6 +3,7 @@ package io.islandtime.ranges
 import io.islandtime.*
 import io.islandtime.Time.Companion.MIDNIGHT
 import io.islandtime.measures.*
+import io.islandtime.operators.startOfDayAt
 import io.islandtime.test.AbstractIslandTimeTest
 import kotlin.test.*
